requirements.txt
ago==0.0.93
attrs==19.3.0
Automat==20.2.0
awscli==1.18.56
backcall==0.1.0
beautifulsoup4==4.9.0
bleach @ file:///home/conda/feedstock_root/build_artifacts/bleach_1588608214987/work
botocore==1.16.6
bs4==0.0.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
dotmap==1.3.14
elasticsearch==7.6.0
entrypoints==0.3
feedfinder2==0.0.4
feedparser==5.2.1
hjson==3.0.1
hurry.filesize==0.9
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.6.0
incremental==17.5.0
ipykernel==5.2.1
ipython @ file:///D:/bld/ipython_1588363073263/work
ipython-genutils==0.2.0
jedi==0.17.0
jieba3k==0.35.1
Jinja2==2.11.2
jmespath==0.9.5
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
kiwisolver==1.2.0
langdetect==1.0.8
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
mkl-service==2.3.0
nbconvert==5.6.1
nbformat==5.0.6
news-please==1.5.3
newspaper3k==0.2.8
nltk==3.5
notebook @ file:///D:/bld/notebook_1588887315004/work
numpy @ file:///D:/bld/numpy_1588605187257/work
packaging==20.1
pandas==1.0.3
pandocfilters==1.4.2
parsel==1.6.0
parso==0.7.0
patsy==0.5.1
pickleshare==0.7.5
Pillow==7.1.2
plac==1.1.3
prometheus-client==0.7.1
prompt-toolkit==3.0.5
Protego==0.1.16
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
Pygments==2.6.1
PyHamcrest==2.0.2
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
queuelib==1.5.0
readability-lxml==0.7.1
regex==2020.5.7
requests==2.23.0
requests-file==1.5.1
rsa==3.4.2
s3transfer==0.3.3
scipy==1.3.1
Scrapy==2.1.0
seaborn==0.10.1
Send2Trash==1.5.0
service-identity==18.1.0
six==1.14.0
soupsieve==2.0
statsmodels==0.11.1
terminado==0.8.3
testpath==0.4.4
tinysegmenter==0.3
tldextract==2.2.2
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
Twisted==20.3.0
urllib3==1.25.9
w3lib==1.21.0
warcio==1.7.3
wcwidth==0.1.9
webencodings==0.5.1
wincertstore==0.2
wordcloud==1.7.0
zipp==3.1.0
zope.interface==5.1.0
